ebook support
This took me a really long time and I'm not too happy with it. Couldn't
find a jekyll plugin that did epub publishing that works. So I created
a gulp workflow that creates a temporary folder of processed markdown
files and runs pandadoc on it to create a .epub file. Then using
Calibre.app to create a MOBI.