Sticky Close Navigation 😃
A user consideration for sighted mobile users. Scrolling up or down the
TOC to reach the Close TOC buttons on either end of it can be rather
exhausting, as the TOC can be rather tall. When scrolling within the
TOC a Close TOC button is now always "stuck" to the top of the screen
as a friendly shortcut.

Interestingly enough, such a consideration is not needed for screen
reader users. They can always access the single Close Navigation
feature available to them through the Form Controls of the VoiceOver
Rotor.